Update SSE example in wai-extra. Fixes #635
