+++ wai-extra/example/index.html
@@ -1,8 +1,9 @@
 <!doctype html>
 <head>
-  <title>Hello</title>
+  <title>Example</title>
+  <link rel="icon" href="data:;base64,=">
 </head>
-<h1>Testing</h1>
+<h1>Testing Server-Sent Events</h1>
 
 <div id="output"></div>
